Tie in combo, omit make_index + cpdup, fix directory creation
The previous commit missed the poudriere.conf.sample changes, so they
have been included here.

The "combo" functionality (jails info, failures, clear mounts) has been
tied in.

The poudriere Makefile "beforeinstall" target was updated to create all
the directories at once, and the subdirectory beforeinstall targets were
removed.  For some reason, the subdir beforeinstall targets are not
executed on bmake/dragonfly.

The libexec makefile was modified to conditionally build make_index and
cpdup if DRAGONFLY is not defined.  We don't need these tools (cpdup is
in base and make_index is not needed by pkgng).