Fix and improve poudriere ports
Several options were removed from ports.sh:
  -F (ZFS option)
  -f (ZFS option)
  -B (SVN option)

The rsync method (really cpdup) was implemented, and the git method was
reimplemented.  The remaining methods were removed.

A timehack was added to the portstree list, which required a modification
to common.sh.  This timestamp reflects the last modification time of the
ports tree, causd by use of -u or -c options.