Improve run hooks
Change run hooks to augment the environment rather than provide
them arguments.  This is more robust because the argument order is
no longer critical and it gives the data a key.

The hooks previously added to testport weren't compatible with
bdrewery's implementation, so they were reworked.  The ones provided
by bulk and common.sh were reworked as well.  Before this can be
adopted in the trunk, the sample hook scripts need to be written to
work with environment variables instead of arguments.