Ensure that json.awk gets the import it is designed for
The awk script, json.awk, was being fed every file starting with
".poudriere" in the log directory.  By adding additional build data,
which is stored in these ".poudriere." files, the html log page was
inadvertently broken.  Using this glob is way to fragile!  To make it
a bit more robust, feed json.awk script data items that it was
designed for.  In the previous case, it didn't recognized "epoch" so
it encoded it as an object which poudiere.js choked on.  Now it is
much much difficult to feed it data accidentally (the variable name
would have to start with ".ports." or "stat" to do it).