Attempt to fix bulk all check failure, spurious testport failure
Dependency checks for entire bulk build failed, as did a testport run
to check a large port.  The first resulted in errors like "echo: can't
write to stdout" while the testport appeared fail at dependency
installation.  I believe both were caused by the limiting of the size
of the reference image.

It was believed the reference wasn't actually used, but the testport
function doesn't use a slave, it builds directly on the reference.
Rather than determine what the proper size is, just leave it unlimited,
for now.