Generate category makefiles after ports creation / update
Poudriere requires the top level and category makefiles in the ports tree
to calculate dependencies.  The makefiles that dports has are not suitable
(they are dynamic) and they will also be removed in the near future.  As
such, poudriere needs to generate makefiles in the format it wants.  This
works because every port in the ports tree is considered valid -- and thus
should be tied into the build.