Tweak nav bar, remove useless "Top" link
The "Top" link which inperceptably jumps to the top of the page was 
removed.  Previously the jump links on the status header was removed, so
this could be considered part of that change.

Styles were added to the nav links to space them better, "../" was
renamed to "Up" for aesthetic reasons.