Implement fully atomic package repository updating.
This builds on the previous commit which runs the build in a shadow
directory of hardlinks.

The new format entails moving all top-level package files and dirs,
such as repo.txz, All/ and Latest, to a .real_TS directory.

There is a .latest symlink that just points to the .real_TS directory.

Then all top-level files/dirs point to .latest/FILE

This allows the completion of package building to just update the .latest
symlink to point to the new repository location atomically.

On first run, the repository will be updated to the new format.