Add new combo feature: Delete logs older than <days>
The form of the command is "poudriere combo -j <jail> -p <portstree>
  -L <days>".

This command will remove all bulk runs that are older than the specified
number of days and it will also remove all the "latest-per-pkg" logs
in both the jail-tree directory and the global latest-per-pkg directory
above it.  Consequently empty directories are also removed.