Post-trunkmerge cleanup
There were a lot of files that didn't get handled properly (maybe as
a result of a fossil=>git conversion.  Unfortunately for me, there were
a lot of commits to the trunk during my sync work, so now I need to go
back and try to capture those.  That was done in this commit as well.