Fix DataTables error and go back to logo.jpg
The logo.svg is relatively huge (164k) and seems to require something
special as it doesn't display and just downloads if you try to display
in a separate tab.  Maybe the webserver needs a dedicated module, dunno.
In an case, let's go back to the tried and true (and low common
denominator) jpeg version of the logo.

Also, modified poudriere.js to not try to insert empty rows at the end
of a table.  The old DataTables tolerated it, but the new version does
not like it.