More colorization tweaking
Since "Renovating" isn't a phase, don't use the phase color. Let's try
white instead for DF jails.

For slaves, the "light blue" and "light red" were still pretty jarring
so let's give the common 1-4 slaves pastels and move light red and light
blue to the second group.  Also move the pleasant light cyan to the
second group because it's highly used as the "port" color so this will
lesson ambiguity IMO.  Adjust the dark colors 8-14 to match the new
order.