Restore totals at end of siginfo list
The DF poudriere siginfo handling has always had the build summary
after the builder list.  This is because we've had 60-96 builders
in the past and the totals scrolled off the console immediately.

While here, get rid of the "show-log" function.  It's mildly useful,
but it wrappes on two lines and it's aesthetically unpleasing.  The
log file location can be determined by the jail name and timestamp,
both of which are provided with siginfo, so I think the loss is less
than the gain here.