Create analog to "was_a_bulk_run" for testport called "testport_run"
There are several places where common.sh checks to see if the script
running is testport.sh, and the check is recreated every time.  Create
a function for this check, and then use it to prevent html reports
getting generated for testport.sh.  The raw logs are enough and the html
report malfunctions due to missing data.