Swap Skipped and Ignored in build summary
This changes it to the same order as trunk and it really is seems to
be better, especially with regards to color.  But also "skipped" is
more interesting than "ignored" so it should go first.

The original reason for the previous order was to match the web
interface, but there's no requirement for console and web interface
to match summary order.